Workaround TFVC delete branch bug
TFVC has an bug that if you delete a branched folder as part of a wider changeset that leaves some files in a different branched folder, then the branching information is not updated to indicate the branch was deleted even though the folder no longer exists.
If a same named folder is subsequently re-branched the TFVC history will report that the new branched folder is a continuation of the old branched folder, even though a new branch information record will be created.